backend/app/pyproject.toml
[project]
name = "qiushuiai-backend"
version = "0.1.0"
description = ""
requires-python = ">=3.10,<4.0"
dependencies = [
    "fastapi[standard]<1.0.0,>=0.114.2",
    "python-multipart<1.0.0,>=0.0.7",
    "email-validator>=2.1.0.post1,<3.0.0.0",
    "passlib[bcrypt]<2.0.0,>=1.7.4",
    "tenacity<9.0.0,>=8.2.3",
    "pydantic[email]>2.0",
    "emails<1.0,>=0.6",
    "jinja2<4.0.0,>=3.1.4",
    "alembic<2.0.0,>=1.12.1",
    "httpx[socks]>=0.25.1,<1.0.0",
    "psycopg[binary]<4.0.0,>=3.1.13",
    "sqlmodel<1.0.0,>=0.0.21",
    # Pin bcrypt until passlib supports the latest
    "bcrypt==4.0.1",
    "pydantic-settings<3.0.0,>=2.2.1",
    "sentry-sdk[fastapi]<2.0.0,>=1.40.6",
    "pyjwt>=2.8.0",
    # AI 相关依赖项 - 从 main_agent 合并
    "langchain-openai>=0.2.1",
    "langchain-anthropic>=0.2.1",
    "langchain>=0.3.1",
    "openai>=1.51.0",
    "langchain-community>=0.3.1",
    "copilotkit>=0.1.58",
    "uvicorn>=0.31.0",
    "python-dotenv>=1.0.1",
    "langchain-core>=0.3.25",
    "langgraph-cli[inmem]>=0.3.0,<0.4.0",
    "langgraph==0.6.6",
    "langchain-mcp-adapters>=0.0.9",
    "fastmcp>=0.4.1",
    "langchain-community ~=0.3.16",
    "duckduckgo-search>=7.3.0",
    "langgraph-checkpoint-postgres==2.0.21",
    # 向量存储相关依赖
    "langchain-postgres>=0.0.2",
    "numpy>=1.24.0",
    "python-docx>=1.2.0",
    "pypdf>=5.8.0",
    "unstructured>=0.18.9",
    "docling>=1.0.0",
    "dashscope>=1.24.1",
    "oss2>=2.19.1",
    "mistune>=3.1.4",
    "requests>=2.32.3",
]

[tool.uv]
dev-dependencies = [
    "pytest<8.0.0,>=7.4.3",
    "mypy<2.0.0,>=1.8.0",
    "ruff<1.0.0,>=0.2.2",
    "pre-commit<4.0.0,>=3.6.2",
    "types-passlib<2.0.0.0,>=1.7.7.20240106",
    "coverage<8.0.0,>=7.4.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["qiushuiai"]

[tool.mypy]
strict = true
exclude = ["venv", ".venv", "alembic"]

[tool.ruff]
target-version = "py310"
exclude = ["alembic"]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG001", # unused arguments in functions
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "W191",  # indentation contains tabs
    "B904",  # Allow raising exceptions without from e, for HTTPException
]

[tool.ruff.lint.pyupgrade]
# Preserve types, even if a file imports `from __future__ import annotations`.
keep-runtime-typing = true
